+++ CHANGELOG.md
@@ -5,6 +5,7 @@ FEATURES:
 - Support Faithfulness Guard with OpenAI credentials
 - Add name to Registered Model version
 - Support directories and relative file paths for Application Sources
+- Support LLM Settings and Vector Database Settings for LLM Blueprints
 
 ## 0.0.23